Create Caesar_Cipher_ main
Project of Day008 of #100DaystoCode challenge. This code is a caesar cipher where the message is encrypted by the index number in the alphabet according to shift amount inputted by user. 
For instance : If message is "abcde" and shift amount is "5", the message becomes "fghij". Reverse logic for decryption, where "fghij" becomes "abcde" when the shift amount is 5.