Create main.py
Day014 of #100DaysToCode Challenge. On day 14 I have built a high low guessing game, where the user gets to guess whoever has the highest Instagram followers. The name, the description, the country gets printed to provide user info and asked to make a choice. The game goes on as long as the user makes the right guess.