Create main.py
Day025 of #100DaystoCode Challenge. 
This is a US State guessing game where you see a map of the united states ( without the names of the states) and try to guess the names of them. If you guess correctly, the name of the state gets written on top of the co-related place in the map.
At the end of the game, it gives you a CSV file for the states you missed.