Create main.py
Day028 of #100DaystoCodeChallenge. This is a Pomodoro timer program that counts 25 minutes of study time, and then a 5 minute break  until reset button is pressed. Each time 25 minute mark is completed, user gets a green check on the screen. For  every 4th session of study completed, there is a long break of 15 minutes instead of 5 minutes.