Create main.py
Day032 of #100DaystoCode challenge. 
This is a automated birthday wisher program that sends an e-mail to a person on your birthday-list  on their birthday automatically by picking one of the template birthday wish letters and putting their name on to it in corresponding place of the letter.
You can run this code on pythonanywhere on a daily basis and it will run once everyday, and send e-mails if anyone's birthday is that day.