Create main.py
Day033 of #100DaystoCode. This is a block of code that checks the coordinates of the International Space Station(ISS) and compares it to pre-defined coordinates to determine if it is close enough to see it at the sky.
If the ISS is close by and the sun is down (in other words, its dark enough to see it), the program sends a notification e-mail, telling the recipient to look up to the sky.
The coordinates of the ISS is collected with a get request to ISS API and it runs every 60 seconds in a while loop.