Create main.py
Day034 of #100Daysto Code Challenge. This is an updated version of Quiz Program that I did at Day 17. Compared to the old one, the new one is now displayed with a User Interface instead of command line, you can answer by pressing buttons instead of typing your answers in. And the questions are retrieved by API, so that each time the program is executed the set of questions are different.