Create main.py
Day035 of #100DaystoCode Challenge. 
This is a weather notification program, which alerts the user to carry an umbrella, by sending an SMS message if it is going to rain or snow in the next 12 hours.
For SMS service Twilio is used (free account).
You can put this piece of code after editing the commented places on pythonanywhere.com and schedule the code to run every morning before you head out, so you are always prepared for the weather ahead!