Create main.py
Day038 of #100DaystoCode Challenge. This is a Exercise/Calorie Logger/Tracker program that allows user to enter any type of text  such as "Ran 3kms." , "Swam 500meters." "Drove a car 50kms." and the program will convert it in a standardized format, calculating the calories burned during the activity based on the input and enter this data into a new row in user's Google Sheet. For more information , see the example result in result.png in Day038 folder.