Create main.py
Day039 and Day040 of #100DaystoCode challenge. 
This is a flight  search program that searches through every available flight for the next 6 months using Kiwi API.
The API is given a source, destination, beginning date, and ending date parameters and returns the flight search results. 
Destination and lowest price parameters is gathered from Google Sheets by using Sheety API.
If there is a cheaper flight within the result than the price specified in Google Sheets for that specific destination, the program sends an SMS to 1 person, and an e-mail to list of users registered for the flight search program.  The program gets the registered users e-mail information from a different tab in the same google sheet.