Create main.py
Day046 of #100DaystoCode Challenge. This is an automated Spotify Playlist Creator. The Spotify Playlist Creator asks user to enter a date. The on the date that user entered, the program checks billboard.com for the first 100 songs on the billboard for that specific date. 
The first 100 songs gets added to a list to search in Spotify. 
For the songs that are available on Spotify, a new playlist is created for the user,  and the songs that are available are added to the new playlist.