Create main.py
Day051 of #100DaystoCode Challenge. This is a Internet Speed Twitter Bot that measures the download/upload speed using speedtest.net.

Then the bot tweets the speed test results along with the guaranteed speeds by ISP in the contract. Tweet also includes the tag of ISP's twitter.  ISP guaranteed speeds that are in the contract defined as constant variables at the top.