Create main.py
Day052 of #100DaystoCode Challenge. This is an Instagram follower bot, where it follows the entire set of followers of an account that user inputs. 
The idea is that if you have a similar Instagram page ( Let's assume you have an Instagram account related to cooking), you might want to follow people who are already following Instagram pages related to cooking. Once they see you followed them, there is a high chance that they will follow you back, because they are interested in cooking related content.