Add files via upload
Day064 of #100DaystoCode Challenge. This is a Movie rating website, where the user can enter a movie name and then the application will search the movie database using TMDB(The Movie Database) it's API, and present user with the search results to select from .
Once user selects the correct movie , the movie is then added with its picture, title, release year, and its description to the webpage.  User is redirected to a page to enter a review and rating for the movie they have added. Then the webpage automatically ranks the movies it has in its database based on their rating. The users have the capability of updating their review & rating or even delete movies. Built by using (Flask_WTF SQL Alchemy) database.