Add files via upload
Day068 of #100DaystoCode Challenge.  This is a webpage with Login & Logout & Register capability implemented with flask_login and werkzeug.security libraries for authentication. The webpage allows certain URLs to be accessed only if the user is logged in.
Passwords are encrypted with Hash & Salt mechanisms to secure the credentials from brute-force attacks.